(PC-31130)[API] refactor: move venue models to offerers schemas
