+++ api/src/pcapi/core/offerers/api.py
@@ -73,6 +73,7 @@ from pcapi.workers.match_acceslibre_job import match_acceslibre_job
 from . import exceptions
 from . import models
 from . import repository as offerers_repository
+from . import schemas as offerers_schemas
 from . import validation
 
 
@@ -102,7 +103,7 @@ def update_venue(
     modifications: dict,
     author: users_models.User,
     opening_hours: list[serialize_base.OpeningHoursModel] | None = None,
-    contact_data: serialize_base.VenueContactModel | None = None,
+    contact_data: offerers_schemas.VenueContactModel | None = None,
     criteria: list[criteria_models.Criterion] | offerers_constants.T_UNCHANGED = offerers_constants.UNCHANGED,
     external_accessibility_url: str | None | offerers_constants.T_UNCHANGED = offerers_constants.UNCHANGED,
     is_manual_edition: bool = False,
@@ -338,7 +339,7 @@ def update_venue_collective_data(
     return venue
 
 
-def upsert_venue_contact(venue: models.Venue, contact_data: serialize_base.VenueContactModel) -> models.Venue:
+def upsert_venue_contact(venue: models.Venue, contact_data: offerers_schemas.VenueContactModel) -> models.Venue:
     """
     Create and attach a VenueContact to a Venue if it has none.
     Update (replace) an existing VenueContact otherwise.
